AbstractMethodUnitTest: take advantage of the change in reportWidth handling
For the tests using the `AbstractMethodUnitTest` class, the `reportWidth` and most other config settings are irrelevant.

This commit changes some of the set up/tear down for the test to make the use of the `Config` class more efficient.

This should make the tests using the `AbstractMethodUnitTest` class as their base significantly faster (at the very least on Windows).

While not benchmarked properly, I have done some comparisons with the test runs on my local machine on Windows.

* `phpunit --filter Core` (= the tests which use this base class + a few extra tests):
    Before: **2 minutes**.
    After: **8 seconds**.
* The same effect can be seen when running `phpunit` without a `--filter`:
    Before: **7 minutes**.
    After: **5 minutes**.
* And when I apply a similar change to the one made here to the base test class in PHPCSUtils (4621 tests):
    Before: **2.5 minutes**.
    After: **1 second**.